requirements.txt
Django==2.2.24
pip==20.1.1
pytz==2021.3
setuptools==47.1.0
sqlparse==0.4.2
psycopg2-binary==2.8.6
